# back-end/modules/employee-service/src/main/resources/static/openapi/employee-api.yml
openapi: 3.0.3
info:
    title: Employee Service API
    version: v1
    description: API quản lý nhân viên, phòng ban, chức danh, thông tin cá nhân và trạng thái làm việc.
servers:
    - url: /
      description: Server đang phục vụ Swagger UI
security:
    - bearerAuth: []
...
paths:
...
    /api/v1/departments:
        post:
            tags: [Departments]
            operationId: createDepartment
            summary: Tạo phòng ban
            description: Mã phòng ban phải duy nhất (phân biệt hoa/thường). Chuỗi được bỏ khoảng trắng đầu/cuối. Phòng ban mới mặc định active=true.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DepartmentRequest"
                        example:
                            code: IT
                            name: Information Technology
                            description: Quản lý hệ thống công nghệ thông tin
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "201":
                    description: Đã tạo phòng ban
                    headers:
                        Location:
                            description: Đường dẫn chi tiết phòng ban vừa tạo
                            schema:
                                type: string
                                example: /api/v1/departments/22222222-2222-2222-2222-222222222222
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartmentResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "409":
                    $ref: "#/components/responses/Conflict"
        get:
            tags: [Departments]
            operationId: listDepartments
            summary: Danh sách phòng ban
            description: Phân trang, sắp xếp theo mã phòng ban tăng dần.
            parameters:
                - name: page
                  in: query
                  description: Số trang, bắt đầu từ 0
                  schema:
                      type: integer
                      format: int32
                      minimum: 0
                      default: 0
                - name: size
                  in: query
                  description: Số phòng ban mỗi trang
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 20
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Danh sách phòng ban và thông tin phân trang
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartmentPage"
                "400":
                    $ref: "#/components/responses/BadRequest"
    /api/v1/departments/{id}:
        get:
            tags: [Departments]
            operationId: getDepartment
            summary: Chi tiết phòng ban
            parameters:
                - name: id
                  in: path
                  required: true
                  description: UUID của phòng ban
                  schema:
                      type: string
                      format: uuid
                  example: 22222222-2222-2222-2222-222222222222
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Thông tin phòng ban
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartmentResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"

        put:
            tags: [Departments]
            operationId: updateDepartment
            summary: Cập nhật phòng ban
            description: >-
                Thay thế code, name và description. Phải gửi đủ code và name;
                description bị bỏ qua hoặc gửi null sẽ được xóa. Mã phải duy nhất
                giữa các phòng ban, được phép giữ mã hiện tại. Chuỗi được bỏ khoảng
                trắng đầu/cuối. Giữ nguyên id, createdAt và active; tự cập nhật
                updatedAt khi dữ liệu thay đổi.
            parameters:
                - name: id
                  in: path
                  required: true
                  description: UUID của phòng ban
                  schema:
                      type: string
                      format: uuid
                  example: 22222222-2222-2222-2222-222222222222
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DepartmentRequest"
                        example:
                            code: IT
                            name: Information Technology
                            description: Phòng công nghệ thông tin
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Thông tin phòng ban sau cập nhật
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DepartmentResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "409":
                    $ref: "#/components/responses/Conflict"
    /api/v1/positions:
        post:
            tags: [Positions]
            operationId: createPosition
            summary: Tạo chức danh
            description: Mã chức danh phải duy nhất (phân biệt hoa/thường). Chuỗi được bỏ khoảng trắng đầu/cuối. Chức danh mới mặc định active=true.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PositionRequest"
                        example:
                            code: SOFTWARE_ENGINEER
                            name: Software Engineer
                            description: Phát triển phần mềm
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "201":
                    description: Đã tạo chức danh
                    headers:
                        Location:
                            description: Đường dẫn chi tiết chức danh vừa tạo
                            schema:
                                type: string
                                example: /api/v1/positions/aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaa2
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PositionResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "409":
                    $ref: "#/components/responses/Conflict"
        get:
            tags: [Positions]
            operationId: listPositions
            summary: Danh sách chức danh
            description: Phân trang, sắp xếp theo mã chức danh tăng dần.
            parameters:
                - name: page
                  in: query
                  description: Số trang, bắt đầu từ 0
                  schema:
                      type: integer
                      format: int32
                      minimum: 0
                      default: 0
                - name: size
                  in: query
                  description: Số chức danh mỗi trang
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 20
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Danh sách chức danh và thông tin phân trang
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PositionPage"
                "400":
                    $ref: "#/components/responses/BadRequest"
    /api/v1/positions/{id}:
        get:
            tags: [Positions]
            operationId: getPosition
            summary: Chi tiết chức danh
            parameters:
                - name: id
                  in: path
                  required: true
                  description: UUID của chức danh
                  schema:
                      type: string
                      format: uuid
                  example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaa2
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Thông tin chức danh
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PositionResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"

        put:
            tags: [Positions]
            operationId: updatePosition
            summary: Cập nhật chức danh
            description: >-
                Thay thế code, name và description. Phải gửi đủ code và name;
                description bị bỏ qua hoặc gửi null sẽ được xóa. Mã phải duy nhất
                giữa các chức danh, được phép giữ mã hiện tại. Chuỗi được bỏ khoảng
                trắng đầu/cuối. Giữ nguyên id, createdAt và active; tự cập nhật
                updatedAt khi dữ liệu thay đổi.
            parameters:
                - name: id
                  in: path
                  required: true
                  description: UUID của chức danh
                  schema:
                      type: string
                      format: uuid
                  example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaa2
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PositionRequest"
                        example:
                            code: SOFTWARE_ENGINEER
                            name: Software Engineer
                            description: Kỹ sư phần mềm
            responses:
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "200":
                    description: Thông tin chức danh sau cập nhật
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PositionResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "409":
                    $ref: "#/components/responses/Conflict"
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Access token RS256 từ Auth; không dùng refresh token.
...
    responses:
        Unauthorized:
            description: Thiếu access token hoặc token sai chữ ký, issuer, audience, thời hạn hay claim bắt buộc.
            headers:
                WWW-Authenticate:
                    schema:
                        type: string
                    example: Bearer
            content:
                application/problem+json:
                    schema:
                        type: object
                        properties:
                            type:
                                type: string
                            title:
                                type: string
                            status:
                                type: integer
                            detail:
                                type: string
                    example:
                        type: about:blank
                        title: Unauthorized
                        status: 401
                        detail: A valid access token is required
...
        BadRequest:
            description: Dữ liệu không hợp lệ, sai UUID/phân trang hoặc tạo vòng lặp quản lý
            content:
                application/problem+json:
                    schema:
                        $ref: "#/components/schemas/ProblemDetail"
                    example:
                        type: about:blank
                        title: Bad Request
                        status: 400
                        detail: Request validation failed
                        instance: /api/v1/employees
                        errors:
                            employeeCode: must not be blank
        NotFound:
            description: Nhân viên, phòng ban, chức danh hoặc department/position/manager được tham chiếu không tồn tại
            content:
                application/problem+json:
                    schema:
                        $ref: "#/components/schemas/ProblemDetail"
        Conflict:
            description: Trùng mã chức danh, mã phòng ban, mã nhân viên/email hoặc xung đột ràng buộc dữ liệu
            content:
                application/problem+json:
                    schema:
                        $ref: "#/components/schemas/ProblemDetail"
    schemas:
        DepartmentRequest:
            type: object
            required: [code, name]
            properties:
                code:
                    type: string
                    minLength: 1
                    maxLength: 50
                    description: Mã phòng ban duy nhất, không được chỉ chứa khoảng trắng
                    example: IT
                name:
                    type: string
                    minLength: 1
                    maxLength: 150
                    description: Tên phòng ban, không được chỉ chứa khoảng trắng
                    example: Information Technology
                description:
                    type: string
                    nullable: true
                    description: Mô tả tùy chọn
        DepartmentResponse:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                code:
                    type: string
                    example: IT
                name:
                    type: string
                    example: Information Technology
                description:
                    type: string
                    nullable: true
        DepartmentPage:
            type: object
            properties:
                content:
                    type: array
                    items:
                        $ref: "#/components/schemas/DepartmentResponse"
                page:
                    type: integer
                    format: int32
                    example: 0
                size:
                    type: integer
                    format: int32
                    example: 20
                totalElements:
                    type: integer
                    format: int64
                totalPages:
                    type: integer
                    format: int32
        PositionRequest:
            type: object
            required: [code, name]
            properties:
                code:
                    type: string
                    minLength: 1
                    maxLength: 50
                    description: Mã chức danh duy nhất, không được chỉ chứa khoảng trắng
                    example: SOFTWARE_ENGINEER
                name:
                    type: string
                    minLength: 1
                    maxLength: 150
                    description: Tên chức danh, không được chỉ chứa khoảng trắng
                    example: Software Engineer
                description:
                    type: string
                    nullable: true
                    description: Mô tả tùy chọn
        PositionResponse:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                code:
                    type: string
                    example: SOFTWARE_ENGINEER
                name:
                    type: string
                    example: Software Engineer
                description:
                    type: string
                    nullable: true
        PositionPage:
            type: object
            properties:
                content:
                    type: array
                    items:
                        $ref: "#/components/schemas/PositionResponse"
                page:
                    type: integer
                    format: int32
                    example: 0
                size:
                    type: integer
                    format: int32
                    example: 20
                totalElements:
                    type: integer
                    format: int64
                totalPages:
                    type: integer
                    format: int32
...
        ProblemDetail:
            type: object
            properties:
                type:
                    type: string
                    format: uri-reference
                    example: about:blank
                title:
                    type: string
                status:
                    type: integer
                    format: int32
                detail:
                    type: string
                instance:
                    type: string
                    format: uri-reference
                errors:
                    type: object
                    description: Thông báo theo tên trường khi validation body thất bại
                    additionalProperties:
                        type: string